create fixing for db, create trigger for Hotel and hotel_reviews

+++ Realta.Contract/Models/v1/Hotels/HotelsDto.cs
@@ -22,7 +22,7 @@ namespace Realta.Contract.Models.v1.Hotels
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public string? HotelReasonStatus { get; set; }
         [AllowNull]
-        public short HotelRatingStar { get; set; }
+        public decimal HotelRatingStar { get; set; }
 
         [Phone]
         [Required(ErrorMessage = "Hotel phone number is required")]
